server/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.9.2
python-multipart==0.0.12
